requirements.txt
django
requests
pyzbar
tailwind
django-tailwind[cookiecutter,honcho,reload]
django-apscheduler 
fastapi[standard]
cron_converter